# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShortURL Pro API",
    "description": "ShortURL Pro 是一个基于 Spring Boot 的短链接服务系统，支持链接缩短、访问统计、用户管理等功能。\n\n## 核心功能\n- 🔗 短链接生成：将长URL转换为短码\n- 🔄 302跳转：根据短码重定向到原始链接\n- 📊 访问统计：实时记录点击次数\n- 🔐 用户管理：基于Spring Security的权限控制\n- 🛠️ 管理功能：启用/禁用短链接状态\n\n## 访问权限\n- **公开接口**：短链接生成、跳转、演示页\n- **认证接口**：用户登录、登出\n- **管理员接口**：需要 ROLE_ADMIN 角色",
    "version": "1.0.0",
    "contact": {
      "name": "ShortURL Pro Team",
      "email": "support@shorturlpro.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "开发环境"
    },
    {
      "url": "https://api.shorturlpro.com",
      "description": "生产环境"
    }
  ],
...
  "paths": {
    "/api/short-url/generate": {
      "post": {
        "tags": ["短链接服务"],
        "summary": "生成短链接",
        "description": "将原始长链接转换为短链接\n\n**权限要求**：公开接口，无需认证",
        "operationId": "generateShortUrl",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortUrlGenerateRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "originalUrl": {
                    "type": "string",
                    "description": "原始长链接",
                    "example": "https://www.example.com/very/long/url/that/needs/to/be/shortened"
                  },
                  "appId": {
                    "type": "string",
                    "description": "应用标识（可选）",
                    "example": "myapp"
                  }
                },
                "required": ["originalUrl"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "短链接生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortUrlGenerateResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/{shortCode}": {
      "get": {
        "tags": ["短链接服务"],
        "summary": "短链接跳转",
        "description": "根据短码进行302重定向到原始链接，并更新访问统计\n\n**权限要求**：公开接口，无需认证",
        "operationId": "redirectToOriginalUrl",
        "parameters": [
          {
            "name": "shortCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 20,
              "pattern": "^[a-zA-Z0-9]+$"
            },
            "description": "短链接码",
            "example": "abc123xyz"
          }
        ],
        "responses": {
          "302": {
            "description": "重定向到原始链接",
            "headers": {
              "Location": {
                "description": "原始链接地址",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          },
          "404": {
            "description": "短链接不存在或已被禁用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/short-url": {
      "post": {
        "tags": ["管理接口"],
        "summary": "创建短链接",
        "description": "创建新的短链接记录\n\n**权限要求**：需要认证",
        "operationId": "createShortUrl",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortUrlCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/short-url/{id}": {
      "put": {
        "tags": ["管理接口"],
        "summary": "更新短链接",
        "description": "根据ID更新短链接信息\n\n**权限要求**：需要认证",
        "operationId": "updateShortUrl",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "短链接ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortUrlCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "短链接不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["管理接口"],
        "summary": "删除短链接",
        "description": "根据ID删除短链接记录\n\n**权限要求**：需要认证",
        "operationId": "deleteShortUrl",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "短链接ID"
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "短链接不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/short-url/status/{id}": {
      "patch": {
        "tags": ["管理接口"],
        "summary": "更新短链接状态",
        "description": "更新短链接的状态（启用/禁用）\n\n**权限要求**：需要认证",
        "operationId": "updateShortUrlStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "短链接ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["ENABLED", "DISABLED"],
                    "description": "目标状态"
                  }
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "状态更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "状态参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "短链接不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/short-url/stats": {
      "get": {
        "tags": ["管理接口"],
        "summary": "获取系统统计信息",
        "description": "获取系统的总体统计信息\n\n**权限要求**：需要认证",
        "operationId": "getSystemStats",
        "responses": {
          "200": {
            "description": "统计信息获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalLinks": {
                      "type": "integer",
                      "description": "总短链接数"
                    },
                    "activeLinks": {
                      "type": "integer",
                      "description": "启用的短链接数"
                    },
                    "totalClicks": {
                      "type": "integer",
                      "description": "总点击次数"
                    },
                    "todayClicks": {
                      "type": "integer",
                      "description": "今日点击次数"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "get": {
        "tags": ["认证接口"],
        "summary": "登录页面",
        "description": "显示登录页面\n\n**权限要求**：公开接口",
        "operationId": "showLoginPage",
        "responses": {
          "200": {
            "description": "登录页面",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["认证接口"],
        "summary": "用户登录",
        "description": "用户登录认证\n\n**权限要求**：公开接口",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "用户名"
                  },
                  "password": {
                    "type": "string",
                    "description": "密码"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "登录成功，重定向到管理页面"
          },
          "401": {
            "description": "登录失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ShortUrlGenerateRequest": {
        "type": "object",
        "required": ["originalUrl"],
        "properties": {
          "originalUrl": {
            "type": "string",
            "format": "uri",
            "description": "原始长链接",
            "example": "https://www.example.com/very/long/url/that/needs/to/be/shortened"
          },
          "appId": {
            "type": "string",
            "description": "应用标识（可选）",
            "example": "myapp",
            "nullable": true
          }
        }
      },
      "ShortUrlCreateRequest": {
        "type": "object",
        "required": ["originalUrl", "name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "短链接名称",
            "example": "我的网站",
            "maxLength": 100
          },
          "originalUrl": {
            "type": "string",
            "format": "uri",
            "description": "原始长链接",
            "example": "https://www.example.com"
          },
          "appId": {
            "type": "string",
            "description": "应用标识（可选）",
            "example": "myapp",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "过期时间（可选）",
            "example": "2026-12-31T23:59:59Z",
            "nullable": true
          }
        }
      },
      "ShortUrlGenerateResponse": {
        "type": "object",
        "required": ["shortCode", "shortUrl"],
        "properties": {
          "shortCode": {
            "type": "string",
            "description": "生成的短码",
            "example": "abc123xyz"
          },
          "shortUrl": {
            "type": "string",
            "format": "uri",
            "description": "完整的短链接",
            "example": "http://localhost:8080/abc123xyz"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "错误发生时间"
          },
          "status": {
            "type": "integer",
            "description": "HTTP状态码"
          },
          "error": {
            "type": "string",
            "description": "错误类型"
          },
          "message": {
            "type": "string",
            "description": "错误消息"
          },
          "path": {
            "type": "string",
            "description": "请求路径"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
}